requirements.txt
browser-use
langchain-anthropic
langchain-openai
langchain-ollama
python-dotenv
fastapi
uvicorn[standard]
supabase
pydantic
requests
